Split into two different apps per comments
