+++ _includes/tutorials/multiple-event-types/confluent/code/configuration/dev.properties
@@ -8,9 +8,9 @@ key.serializer=org.apache.kafka.common.serialization.StringSerializer
 # Application specific properties
 proto.topic.name=proto-events
 proto.topic.partitions=1
-proto.topic.replication.factor=1
+proto.topic.replication.factor=3
 
 avro.topic.name=avro-events
 avro.topic.partitions=1
-avro.topic.replication.factor=1
+avro.topic.replication.factor=3
