+++ _includes/tutorials/multiple-event-types/kafka/code/Makefile
@@ -6,5 +6,6 @@ tutorial:
 	rm -r $(DEV_OUTPUTS_DIR) || true
 	mkdir $(DEV_OUTPUTS_DIR)
 	harness-runner ../../../../../_data/harnesses/multiple-event-types/kafka.yml $(TEMP_DIR)
-	diff --strip-trailing-cr $(STEPS_DIR)/dev/expected-output.txt $(DEV_OUTPUTS_DIR)/actual-output.txt
+	diff --strip-trailing-cr $(STEPS_DIR)/dev/expected-output-protobuf.txt $(DEV_OUTPUTS_DIR)/actual-output-protobuf.txt
+	diff --strip-trailing-cr $(STEPS_DIR)/dev/expected-output-avro.txt $(DEV_OUTPUTS_DIR)/actual-output-avro.txt
 	reset
