+++ _includes/tutorials/multiple-event-types/kafka/code/configuration/test.properties
@@ -1,8 +1,3 @@
-input.topic.name=input-topic
-input.topic.partitions=1
-input.topic.replication.factor=1
-
-
 proto.topic=proto-records
 avro.topic=avro-records
 schema.registry.url=mock://multi-event-produce-consume-test
