+++ _includes/tutorials/multiple-event-types/kafka/code/src/main/java/io/confluent/developer/MultiEventAvroProduceConsumeApp.java
@@ -2,16 +2,10 @@ package io.confluent.developer;
 
 import io.confluent.developer.avro.PageView;
 import io.confluent.developer.avro.Purchase;
-import io.confluent.developer.proto.CustomerEventProto;
-import io.confluent.developer.proto.PageViewProto;
-import io.confluent.developer.proto.PurchaseProto;
 import io.confluent.kafka.serializers.AbstractKafkaSchemaSerDeConfig;
 import io.confluent.kafka.serializers.KafkaAvroDeserializer;
 import io.confluent.kafka.serializers.KafkaAvroDeserializerConfig;
 import io.confluent.kafka.serializers.KafkaAvroSerializer;
-import io.confluent.kafka.serializers.protobuf.KafkaProtobufDeserializer;
-import io.confluent.kafka.serializers.protobuf.KafkaProtobufDeserializerConfig;
-import io.confluent.kafka.serializers.protobuf.KafkaProtobufSerializer;
 import org.apache.avro.specific.SpecificRecord;
 import org.apache.avro.specific.SpecificRecordBase;
 import org.apache.kafka.clients.admin.AdminClient;
@@ -40,30 +34,12 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.function.Supplier;
 
-public class MultiEventProduceConsumeApp implements AutoCloseable{
+public class MultiEventAvroProduceConsumeApp implements AutoCloseable{
 
     public static final String CUSTOMER_ID = "wilecoyote";
-    private static final Logger LOG = LoggerFactory.getLogger(MultiEventProduceConsumeApp.class);
+    private static final Logger LOG = LoggerFactory.getLogger(MultiEventAvroProduceConsumeApp.class);
     private volatile boolean keepConsumingAvro = true;
-    private volatile boolean keepConsumingProto = true;
-    final ExecutorService executorService = Executors.newFixedThreadPool(2);
-
-    public void produceProtobufEvents(final Supplier<Producer<String, CustomerEventProto.CustomerEvent>> producerSupplier,
-                                      final String topic,
-                                      final List<CustomerEventProto.CustomerEvent> protoCustomerEvents) {
-
-        try (Producer<String, CustomerEventProto.CustomerEvent> producer = producerSupplier.get()) {
-            protoCustomerEvents.stream()
-                    .map((event -> new ProducerRecord<>(topic, event.getId(), event)))
-                    .forEach(producerRecord -> producer.send(producerRecord, ((metadata, exception) -> {
-                        if (exception != null) {
-                            LOG.error("Error Protobuf producing message", exception);
-                        } else {
-                            LOG.debug("Produced Protobuf record offset {} timestamp {}", metadata.offset(), metadata.timestamp());
-                        }
-                    })));
-        }
-    }
+    final ExecutorService executorService = Executors.newFixedThreadPool(1);
 
     public void produceAvroEvents(final Supplier<Producer<String, SpecificRecordBase>> producerSupplier,
                                   final String topic,
@@ -79,31 +55,6 @@ public class MultiEventProduceConsumeApp implements AutoCloseable{
                             LOG.debug("Produced Avro record offset {} timestamp {}", metadata.offset(), metadata.timestamp());
                         }
                     })));
-        }
-    }
-
-    public void consumeProtoEvents(final Supplier<Consumer<String, CustomerEventProto.CustomerEvent>> consumerSupplier,
-                                   final String topic,
-                                   final List<String> eventTracker) {
-
-        try (Consumer<String, CustomerEventProto.CustomerEvent> eventConsumer = consumerSupplier.get()) {
-            eventConsumer.subscribe(Collections.singletonList(topic));
-            while (keepConsumingProto) {
-                ConsumerRecords<String, CustomerEventProto.CustomerEvent> consumerRecords = eventConsumer.poll(Duration.ofSeconds(1));
-                consumerRecords.forEach(consumerRec -> {
-                    CustomerEventProto.CustomerEvent customerEvent = consumerRec.value();
-                    switch (customerEvent.getActionCase()) {
-                        case PURCHASE:
-                            eventTracker.add(String.format("Protobuf Purchase event -> %s", customerEvent.getPurchase().getItem()));
-                            break;
-                        case PAGE_VIEW:
-                            eventTracker.add(String.format("Protobuf Pageview event -> %s", customerEvent.getPageView().getUrl()));
-                            break;
-                        default:
-                            LOG.error("Unexpected type - this shouldn't happen");
-                    }
-                });
-            }
         }
     }
 
@@ -149,28 +100,8 @@ public class MultiEventProduceConsumeApp implements AutoCloseable{
         return events;
     }
 
-    public List<CustomerEventProto.CustomerEvent> protobufEvents() {
-        CustomerEventProto.CustomerEvent.Builder customerEventBuilder = CustomerEventProto.CustomerEvent.newBuilder();
-        PageViewProto.PageView.Builder pageViewBuilder = PageViewProto.PageView.newBuilder();
-        PurchaseProto.Purchase.Builder purchaseBuilder = PurchaseProto.Purchase.newBuilder();
-        List<CustomerEventProto.CustomerEvent> events = new ArrayList<>();
-
-        PageViewProto.PageView pageView = pageViewBuilder.setCustomerId(CUSTOMER_ID).setUrl("http://acme/traps").setIsSpecial(false).build();
-        PageViewProto.PageView pageViewII = pageViewBuilder.setCustomerId(CUSTOMER_ID).setUrl("http://acme/bombs").setIsSpecial(false).build();
-        PageViewProto.PageView pageViewIII = pageViewBuilder.setCustomerId(CUSTOMER_ID).setUrl("http://acme/bait").setIsSpecial(true).build();
-        PurchaseProto.Purchase purchase = purchaseBuilder.setCustomerId(CUSTOMER_ID).setItem("road-runner-bait").setAmount(99.99).build();
-
-        events.add(customerEventBuilder.setId(CUSTOMER_ID).setPageView(pageView).build());
-        events.add(customerEventBuilder.setId(CUSTOMER_ID).setPageView(pageViewII).build());
-        events.add(customerEventBuilder.setId(CUSTOMER_ID).setPageView(pageViewIII).build());
-        events.add((customerEventBuilder.setId(CUSTOMER_ID).setPurchase(purchase)).build());
-
-        return events;
-    }
-
     @Override
     public void close() {
-        keepConsumingProto = false;
         keepConsumingAvro = false;
         executorService.shutdown();
     }
@@ -180,12 +111,7 @@ public class MultiEventProduceConsumeApp implements AutoCloseable{
         try (final AdminClient client = AdminClient.create(allProps)) {
 
             final List<NewTopic> topics = new ArrayList<>();
-
-            topics.add(new NewTopic(
-                    allProps.getProperty("proto.topic.name"),
-                    Integer.parseInt(allProps.getProperty("proto.topic.partitions")),
-                    Short.parseShort(allProps.getProperty("proto.topic.replication.factor"))));
-
+            
             topics.add(new NewTopic(
                     allProps.getProperty("avro.topic.name"),
                     Integer.parseInt(allProps.getProperty("avro.topic.partitions")),
@@ -209,21 +135,9 @@ public class MultiEventProduceConsumeApp implements AutoCloseable{
         properties.forEach((key, value) -> commonConfigs.put((String) key, value));
 
 
-        try (MultiEventProduceConsumeApp multiEventApp = new MultiEventProduceConsumeApp()) {
+        try (MultiEventAvroProduceConsumeApp multiEventApp = new MultiEventAvroProduceConsumeApp()) {
             multiEventApp.createTopics(properties);
-            String protobufTopic = (String) commonConfigs.get("proto.topic.name");
             String avroTopic = (String) commonConfigs.get("avro.topic.name");
-
-            LOG.info("Producing Protobuf events now");
-            multiEventApp.produceProtobufEvents(() -> new KafkaProducer<>(protoProduceConfigs(commonConfigs)), protobufTopic, multiEventApp.protobufEvents());
-
-            List<String> protoEvents = new ArrayList<>();
-            multiEventApp.executorService.submit(() -> multiEventApp.consumeProtoEvents(() -> new KafkaConsumer<>(protoConsumeConfigs(commonConfigs)), protobufTopic, protoEvents));
-            while (protoEvents.size() < 3) {
-                Thread.sleep(100);
-            }
-            LOG.info("Consumed Proto Events {}", protoEvents);
-
 
             LOG.info("Producing Avro events");
             multiEventApp.produceAvroEvents(() -> new KafkaProducer<>(avroProduceConfigs(commonConfigs)), avroTopic, multiEventApp.avroEvents());
@@ -256,21 +170,5 @@ public class MultiEventProduceConsumeApp implements AutoCloseable{
         avroProduceConfigs.put(AbstractKafkaSchemaSerDeConfig.AUTO_REGISTER_SCHEMAS, false);
         avroProduceConfigs.put(AbstractKafkaSchemaSerDeConfig.USE_LATEST_VERSION, true);
         return avroProduceConfigs;
-    }
-
-    @NotNull
-    static Map<String, Object> protoConsumeConfigs(Map<String, Object> commonConfigs) {
-        Map<String, Object> protoConsumeConfigs = new HashMap<>(commonConfigs);
-        protoConsumeConfigs.put(ConsumerConfig.GROUP_ID_CONFIG, "protobuf-consumer-group");
-        protoConsumeConfigs.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, KafkaProtobufDeserializer.class);
-        protoConsumeConfigs.put(KafkaProtobufDeserializerConfig.SPECIFIC_PROTOBUF_VALUE_TYPE, CustomerEventProto.CustomerEvent.class);
-        return protoConsumeConfigs;
-    }
-
-    @NotNull
-    static Map<String, Object> protoProduceConfigs(Map<String, Object> commonConfigs) {
-        Map<String, Object> protoProduceConfigs = new HashMap<>(commonConfigs);
-        protoProduceConfigs.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, KafkaProtobufSerializer.class);
-        return protoProduceConfigs;
     }
 }
