+++ _includes/tutorials/multiple-event-types/kafka/markup/dev/make-multi-event-app-avro.adoc
@@ -1,4 +1,4 @@
-Next up in the tutorial is the code you'll use to create the Avro producer.
+Next up in the tutorial is to create the application for the Avro multiple events application.
 [source, java]
 .KafkaProducer for Avro events
 ----
@@ -25,7 +25,7 @@ public void produceAvroEvents(final Supplier<Producer<String, SpecificRecordBase
 <4> Specifying **_to disable_** automatic schema registation
 <5> Setting to use the latest schema
 
-You have a very similar approach with the Avro producer, but take a took at the type at annotation one - it's an abstract class, `SpecificRecordBase` that every Avro generated class inherits from.  Since the schema for the Avro multi-event topic uses a `union` at the top level, you don't know the concrete type.  Since you want to use the customer id as the key you need to access the field in a map-like fashion by using the field name as it exists in the schema.  This is possible because `SpecificRecordBase` implements the `GenericRecord` interface which provides the `get` method for retrieving a field value by name.
+You have a very similar approach with the Avro producer as you did with the Protobuf version, but take a took at the type at annotation one - it's an abstract class, `SpecificRecordBase` that every Avro generated class inherits from.  Since the schema for the Avro multi-event topic uses a `union` at the top level, you don't know the concrete type.  Since you want to use the customer id as the key you need to access the field in a map-like fashion by using the field name as it exists in the schema.  This is possible because `SpecificRecordBase` implements the `GenericRecord` interface which provides the `get` method for retrieving a field value by name.
 
 But the biggest difference is the configurations you provide to the producer for the Avro serializer, namely disabling automatic schema registration, otherwise it would override the union schema as the latest one.  Additionally since you've set `use.latest.version` to `true` the serializer looks up the latest version, the union schema, and will use that for serialization. https://www.confluent.io/blog/multiple-event-types-in-the-same-kafka-topic/#avro-unions-with-schema-references[This blog post] by Robert Yokota explains this mechanism in detail.
 
@@ -55,25 +55,20 @@ With the Avro consumer you'll need to use the Java `instanceof` operator to dete
 Before you go on to create the application we should mention quickly about the deserialization configurations you need to set
 
 [source, java]
-.Configurations needed by the Consumer for both Avro and Protobuf
+.Configurations needed by the Avro Consumer
 ----
  avroConsumeConfigs.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, KafkaAvroDeserializer.class); <1>
  avroConsumeConfigs.put(KafkaAvroDeserializerConfig.SPECIFIC_AVRO_READER_CONFIG, true); <2>
-
- protoConsumeConfigs.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, KafkaProtobufDeserializer.class); <3>
- protoConsumeConfigs.put(KafkaProtobufDeserializerConfig.SPECIFIC_PROTOBUF_VALUE_TYPE, CustomerEventProto.CustomerEvent.class); <4>
 ----
 
 <1> Specifying to use the Avro deserializer for the Avro consumer
 <2> Setting the Avro deserializer to use the specific reader
-<3> Configurations for the Protobuf consumer to use the Protobuf deserializer
-<4> Setting the specific class type for the Protobuf deserializer
 
-It should come as no surprise that you need to set the deserializer class to either `KafkaAvroDeserializer` and `KafkaProtobufDeserializer` for the Avro and Protobuf consumers.  But when working with multiple types, you still need to set the configuration for a specific type.  For Protobuf it's straight forward, setting the specific type to the outer class makes sense since the proto deserialization process knows how to handle the embedded types due to the schema.  With Avro though even with the `union` schema you'll still need to specify to set the `SPECIFIC_AVRO_READER_CONFIG` to `true` to get the concete types.
+It should come as no surprise that you need to set the deserializer class to `KafkaAvroDeserializer` for the Avro consumer.  But when working with multiple types, you still need to set the configuration for a specific type.  With Avro, even with the `union` schema you'll still need to specify to set the `SPECIFIC_AVRO_READER_CONFIG` to `true` to get the concrete types.
 
 
-Now go ahead and create the `src/main/java/io/confluent/developer/MultiEventProduceConsumeApp.java` file:
+Now go ahead and create the `src/main/java/io/confluent/developer/MultiEventAvroProduceConsumeApp.java` file:
 
 +++++
-<pre class="snippet"><code class="java">{% include_raw tutorials/multiple-event-types/kafka/code/src/main/java/io/confluent/developer/MultiEventProduceConsumeApp.java %}</code></pre>
+<pre class="snippet"><code class="java">{% include_raw tutorials/multiple-event-types/kafka/code/src/main/java/io/confluent/developer/MultiEventAvroProduceConsumeApp.java %}</code></pre>
 +++++
