+++ _includes/tutorials/multiple-event-types/kafka/markup/dev/run-dev-app.adoc
@@ -1,14 +1,30 @@
-Now that you have an uberjar for the KafkaMultiEventConsumerApplication, you can launch it locally. When you run the following, you'll see some output as each producer sends records to the Kafka broker and you'll also see the results of the multiple event consumers.
+Now that you have an uberjar, you can launch each application locally. When you run the following, you'll see some output as the producer sends records to the Kafka broker and you'll also see the results of the multiple event consumer.
 
 NOTE: We'll use a pipe and grep at the end of the command to filter out the logging from the clients which will make the results harder to see
 
+First run the multi-event application for Protobuf
+
 +++++
-<pre class="snippet"><code class="shell">{% include_raw tutorials/multiple-event-types/kafka/code/tutorial-steps/dev/run-dev-app.sh %}</code></pre>
+<pre class="snippet"><code class="shell">{% include_raw tutorials/multiple-event-types/kafka/code/tutorial-steps/dev/run-dev-app-protobuf.sh %}</code></pre>
 +++++
 
 
 The output should look something like this:
 
 +++++
-<pre class="snippet"><code class="shell">{% include_raw tutorials/multiple-event-types/kafka/code/tutorial-steps/dev/expected-output.txt %}</code></pre>
+<pre class="snippet"><code class="shell">{% include_raw tutorials/multiple-event-types/kafka/code/tutorial-steps/dev/expected-output-protobuf.txt %}</code></pre>
++++++
+
+
+Then run the multi-event application for Avro
+
++++++
+<pre class="snippet"><code class="shell">{% include_raw tutorials/multiple-event-types/kafka/code/tutorial-steps/dev/run-dev-app-avro.sh %}</code></pre>
++++++
+
+
+The output should look something like this:
+
++++++
+<pre class="snippet"><code class="shell">{% include_raw tutorials/multiple-event-types/kafka/code/tutorial-steps/dev/expected-output-avro.txt %}</code></pre>
 +++++
