+++ _includes/tutorials/multiple-event-types/kafka/markup/test/invoke-tests.adoc
@@ -1,4 +1,4 @@
-Now run the test, which is as simple as:
+Now run the tests, which is as simple as:
 
 +++++
 <pre class="snippet"><code class="shell">{% include_raw tutorials/multiple-event-types/kafka/code/tutorial-steps/test/invoke-tests.sh %}</code></pre>
